[RAIN-3803] Added logic to enrich answers count in survey search response

+++ core-services/egov-survey-services/src/main/java/org/egov/egovsurveyservices/web/models/SurveyEntity.java
@@ -59,7 +59,7 @@ public class SurveyEntity {
     private Boolean active;
 
     @JsonProperty("answersCount")
-    private Long answersCount;
+    private Integer answersCount;
 
     public SurveyEntity addQuestionsItem(Question questionItem) {
         if (this.questions == null) {
